+++ NEWS.md
@@ -1,9 +1,38 @@
-# immApex VERSION 1.2.1
-* Update unit testing 
-* Removed src .o and .so files
+# immApex VERSION 1.3.2
 
-# immApex VERSION 1.0.6
-* Update unit testing and keras vignette check
+## UNDERLYING CHANGES
+* Rebuilt ```propertyEncoder()``` and ```onehotEncoder()``` to use C++ backend with encodeSequences.cpp
+* Rebuilt ```buildNetwork()``` via C++ integration (fastEditEdges.cpp)
+* Motif quantification in ```calculateMotif()``` use C++ (calculateMotif.cpp)
+* Reduce overall dependencies on external packages
+* Improved speed of ```generateSequences()```, ```mutateSequences()```, ```adjacencyMatrix()```, ```tokenizeSequences()```, ```geometricEncoder```
+* Added group.by argument to ```getIR()```
+* Defunct ```variationalSequences()``` 
+* Removed keras3/tensorflow suggests/dependencies
+* Removed baslisk environment, python no longer required
+* Updated C++ backend to C++17
+
+## NEW FEATURES
+* ```calculateEntropy()``` added to calculate the positional entropy along a biological sequence
+* Added basic diversity metrics (exported) to support ```calculateEntropy()```
+* ```calculateFrequency()``` added to calculate the positional frequency along a biological sequence
+* ```calculateMotif()``` added to get motif quantification of sequences
+* ```calculateGeneUsage()``` added for single/paired gene enumeration
+* Added ```scaleMatrix()``` for comprehensive scale/transformation functions
+* Added ```summaryMatrix()``` for fast summarization of matrix values
+
+## BUGS
+* Fixed NT position issue with ```inferCDR()```
+* Fixed ```generateSequence()``` range issue for single-length sequences
+* Fixed binary frequency issue in ```positionalEncoder()```
+* Fixed ```buildNetwork()``` for relative thresholding returns relative value
+
+# immApex VERSION 1.2.2
+* Updated unit tests and vignette check
+* Converted package to basilisk from reticulate
+
+# immApex VERSION 1.2.0
+* Update version for BioConductor release
 
 # immApex VERSION 1.0.5
 
@@ -15,12 +44,12 @@
 # immApex VERSION 1.0.4
 
 ## UNDERLYING CHANGES
-* Optional testthat variationalSequences() evaulate presence of Keras
+* Optional testthat ```variationalSequences()``` evaluate presence of Keras
 
 # immApex VERSION 1.0.3
 
 ## UNDERLYING CHANGES
-* Drop evaluation of variationalSequences() example
+* Drop evaluation of ```variationalSequences()``` example
 
 # immApex VERSION 1.0.2